api/requirements.txt
apiview==1.3.25
asgiref==3.5.2
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1
django-import-export==2.8.0
django-restframework-apiview==1.4.3
djangorestframework==3.13.1
et-xmlfile==1.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.10
Pillow==9.2.0
pytz==2022.2.1
PyYAML==6.0
reportlab==3.6.11
six==1.16.0
sqlparse==0.4.2
tablib==3.2.1
tzdata==2022.2
xlrd==2.0.1
xlwt==1.3.0
